backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==0.28.1
chromadb==0.4.18
sentence-transformers==2.2.2
numpy<2.0
# boto3==1.34.0  # Commented out - not needed for OpenAI integration

